[feat/#68]: 북마크,관리자 페이지 구조 변경
