--- woori-zip/src/app/manager/layout/Header/Header.module.css
@@ -1,110 +1,0 @@
-.header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 1rem 2rem;
-  background-color: #F4FCFF;
-  font-family: var(--font-Gmarket), sans-serif;
-}
-
-.header_left {
-  margin-right: auto;
-}
-
-.header_title {
-  color: #0065C4;
-  font-size: 2rem;
-  font-weight: bold;
-  text-decoration: none;
-}
-
-.login_button {
-  padding: 0.5rem 1.5rem;
-  background-color: #0065C4;
-  color: white;
-  border: 2px solid #0065C4;
-  border-radius: 30px;
-  font-weight: bold;
-  cursor: pointer;
-  transition: all 0.3s ease;
-}
-
-.login_button:hover {
-  background-color: #0b70cf;
-  color: white;
-}
-
-.webButton {
-  display: block;
-}
-
-.mobileButton {
-  display: none;
-}
-
-.headerNav {
-  width: 100%;
-  max-width: 1000px;
-  margin: 10px auto;
-  padding: 0 20px;
-  background-color: #F4FCFF;
-}
-
-.navList {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  list-style: none;
-  margin: 0;
-  padding: 0;
-  text-align: center;
-}
-
-.navItem {
-  padding: 16px 0;
-  cursor: pointer;
-  border-bottom: 1px solid #eef2f6;
-  color: #666;
-  font-size: 16px;
-}
-
-.navLink {
-  text-decoration: none;
-  color: inherit;
-  font-weight: 500;
-}
-
-.activeItem {
-  color: #0065C4;
-  border-bottom: 2px solid #0065C4;
-}
-
-.navItem:hover {
-  color: #0065C4;
-}
-
-@media screen and (max-width: 393px) {
-  .header {
-    padding: 16px;
-  }
-
-  .header_title {
-    font-size: 1.5rem;
-  }
-
-  .webButton {
-    display: none;
-  }
-
-  .mobileButton {
-    display: block;
-  }
-
-  .headerNav {
-    padding: 0 16px;
-  }
-
-  .navItem {
-    padding: 12px 0;
-    font-size: 14px;
-  }
-}
