+++ woori-zip/src/components/domains/manager/Tabs/Tabs.module.css
