+++ woori-zip/src/components/layout/Footer/Footer.js
@@ -1,6 +1,7 @@
 import styles from './Footer.module.css';
 import Link from 'next/link';
 
+
 export default function Footer() {
  return (
    <div className={styles.footerWrapper}>
